Refactor to make install/uninstall implementation easier
